MSDKUI-1649: Move formatter logic to Maneuver View
This is the first of several commits of an effort to allow developers to
set the formatter and preferred unit used by the GuidanceManeuverView.

This commit moves the formatter from the presenter to the view itself,
like we already have for other views such as the
GuidanceEstimatedArrivalView.

Next commits will allow the developers to intercept the
GuidanceManeuverData and set the unit they want. This way they can use a
formatter which respects the unit passed by the measurement type.

Signed-off-by: Otavio Cordeiro <41054361+otaviohere@users.noreply.github.com>