MSDKUI-1649: Make GuidanceManeuverData properties public
This commit makes all the properties public so developers can intercept
the maneuver data passed by the monitor to the view controller and
change their values if necessary.

This is specially important for the distance property which needs to be
intercepted by developers if they want to force a particular measurement
unit to it.

Signed-off-by: Otavio Cordeiro <41054361+otaviohere@users.noreply.github.com>