MSDKUI-1654: Improve GuidanceManeuverData handling
The GuidanceManeuverData model contains nullable properties and,
therefore, GuidanceManeuverView should reflect that.

Current implementation only updates the view if the values carried by
GuidanceManeuverData are not nil. This creates a problem, since a new
value with nil would keep the previous view value.

This commit changes that. If a new model contains nil, the view will
take that in consideration and reflect whatever the model has.

Signed-off-by: Otavio Cordeiro <41054361+otaviohere@users.noreply.github.com>