MSDKUI-1661: Rename GuidanceCurrentStreetLabel to GuidanceStreetLabel
Since GuidanceCurrentStreetLabel can be used to show any street, not only the
current one, the name of this class should not be associated with it.

The current street data is provided by GuidanceCurrentStreetNameMonitor which
keeps its name.

In GuidanceViewController the GuidanceStreetLabel is paired with
GuidanceCurrentStreetNameMonitor, thus related variable and function names
should keep existing naming.

- renames class name and swift file from GuidanceCurrentStreetLabel to
  GuidanceStreetLabel
- renames class name and swift file from GuidanceCurrentStreetLabelTests to
  GuidanceStreetLabelTests
- updates GuidanceStreetLabel class summary

Signed-off-by: Piotr Marczycki <piotrekm44@users.noreply.github.com>