MSDKUI-1672: Fix Speed Unit for UK
UK uses a mix of Imperial and Metric System. In Swift, Locale has a
property called .usesMetricSystem which is a boolean. For UK, it returns
true.

In Objective-C, Local has 3 possible values for metric system, US, UK,
Metric. This commit uses the Objective-C bridge to check the Locale's
measurement system.

Signed-off-by: Otavio Cordeiro <41054361+otaviohere@users.noreply.github.com>