MSDKUI-1614: Filter Components by presentation style
This commit introduces a new View Controller, ComponentsViewController,
and its data source. After this change 1) it's easier to add new entries
to the table view, since all one needs to do is to add a new
ComponentEntry to the data source, and 2) it's possible to filter by
presentation style (intrinsic, constrained, or all).

Signed-off-by: Otavio Cordeiro <41054361+otaviohere@users.noreply.github.com>